requirements.txt
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.2
urllib3==1.22
Werkzeug==0.14.1
